Split middleware from server.go
